chore: refactor to formatting file
