+++ src/utils/formatting.ts
@@ -47,3 +47,12 @@ export const formatBetHistory = (
   blockNumber: +bet.createdAt,
   marketOracleResultSig: signedBetData.marketOracleResultSig
 });
+
+export const shortenAddress = (address: string) =>
+  `${address.slice(0, 5)}...${address.slice(address.length - 5)}`;
+
+export const shortenHash = (hash: string) => {
+  const start = hash.substring(0, 15);
+  const end = hash.substring(hash.length - 15, hash.length);
+  return `${start}...${end}`;
+};
