--- src/utils/shortenAddress.ts
@@ -1,2 +1,0 @@
-export const shortenAddress = (address: string) =>
-  `${address.slice(0, 5)}...${address.slice(address.length - 5)}`;
