--- src/utils/shortenHash.ts
@@ -1,5 +1,0 @@
-export const shortenHash = (hash: string) => {
-  const start = hash.substring(0, 15);
-  const end = hash.substring(hash.length - 15, hash.length);
-  return `${start}...${end}`;
-};
